Feat: Fixed import error and added status

+++ src/ImpactMarketplace.sol
@@ -2,7 +2,7 @@
 
 pragma solidity ^0.8.20;
 
-import {IERC20} from "lib/openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "./ImpactDAO.sol";
 import "./ImpactRewardToken.sol";
 
@@ -11,7 +11,7 @@ interface SoulNft {
 
     function burn(uint256 tokenId) external;
 
-    function showIds(address _member) external view returns (uint);
+    function showIds(address _member) external view returns (uint256);
 
     function showMembers() external view returns (address[] memory);
 }
@@ -35,18 +35,12 @@ contract ImpactMarketPlace {
     // I'm thinking it should be onlyadmin that can upload and edit or dao members, you decide.
 
     event ProductCreated(
-        uint256 id,
-        string name,
-        string description,
-        uint256 price,
-        string image,
-        uint256 sold,
-        uint256 availableItems
+        uint256 id, string name, string description, uint256 price, string image, uint256 sold, uint256 availableItems
     );
-    event ProductUpdated(uint id, uint256 price, uint256 availableItems);
-    event ProductDeleted(uint id);
-    event ProductBought(uint id, address buyer, uint256 price, uint quantity);
-    event ProductEdited(uint id, uint256 price, uint256 availableItems);
+    event ProductUpdated(uint256 id, uint256 price, uint256 availableItems);
+    event ProductDeleted(uint256 id);
+    event ProductBought(uint256 id, address buyer, uint256 price, uint256 quantity);
+    event ProductEdited(uint256 id, uint256 price, uint256 availableItems);
 
     error NotDAOMember();
 
@@ -78,22 +72,10 @@ contract ImpactMarketPlace {
         p.sold = 0;
         p.availableItems = available;
 
-        emit ProductCreated(
-            id,
-            _name,
-            _description,
-            _price,
-            _image,
-            0,
-            available
-        );
+        emit ProductCreated(id, _name, _description, _price, _image, 0, available);
     }
 
-    function editProduct(
-        uint256 id,
-        uint256 _price,
-        uint256 available
-    ) external {
+    function editProduct(uint256 id, uint256 _price, uint256 available) external {
         if (soulnft.balanceOf(msg.sender) != 1) revert NotDAOMember();
         Product storage p = products[id];
         p.price = _price;
@@ -111,11 +93,8 @@ contract ImpactMarketPlace {
     function buyProduct(uint256 id, uint256 quantity) external payable {
         require(id <= idCounter, "Invalid Product ID");
         Product storage p = products[id];
-        uint quantityPrice = p.price * quantity;
-        require(
-            impactrewards.balanceOf(msg.sender) >= quantityPrice,
-            "Not Enough Impact rewards token"
-        );
+        uint256 quantityPrice = p.price * quantity;
+        require(impactrewards.balanceOf(msg.sender) >= quantityPrice, "Not Enough Impact rewards token");
         require(p.availableItems > 0, "Product not available");
         p.sold++;
         p.availableItems--;
